Exception handling

+++ src/main/java/com/example/devsuperior/dsecomecer/Service/ProductService.java
@@ -1,16 +1,22 @@
 package com.example.devsuperior.dsecomecer.Service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.example.devsuperior.dsecomecer.Service.exceptions.DatabaseException;
+import com.example.devsuperior.dsecomecer.Service.exceptions.ResourceNotFoundExceptions;
 import com.example.devsuperior.dsecomecer.dto.Request.RequestProductDTO;
 import com.example.devsuperior.dsecomecer.dto.Response.ResponseProductDTO;
 import com.example.devsuperior.dsecomecer.entities.Product;
 import com.example.devsuperior.dsecomecer.repositories.ProductRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -23,7 +29,8 @@ public class ProductService {
         // Product product = result.get();
         // ProductDTO productDTO = new ProductDTO(product);
         // return productDTO;
-        Product product = productRepository.findById(id).get();
+        Product product = productRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundExceptions("Produc not found"));
         return new ResponseProductDTO(product);
     }
 
@@ -51,20 +58,33 @@ public class ProductService {
 
     @Transactional
     public ResponseProductDTO update(Long id, RequestProductDTO productDTO) {
-        Product entity = productRepository.getReferenceById(id);
-        copyDtoToEntity(productDTO, entity);
-        // entity.setName(productDTO.getName());
-        // entity.setDescription(productDTO.getDescription());
-        // entity.setImgUrl(productDTO.getImgUrl());
-        // entity.setPrice(productDTO.getPrice());
-        entity = productRepository.save(entity);
-        return new ResponseProductDTO(entity);
+
+        try {
+
+            Product entity = productRepository.getReferenceById(id);
+            copyDtoToEntity(productDTO, entity);
+            // entity.setName(productDTO.getName());
+            // entity.setDescription(productDTO.getDescription());
+            // entity.setImgUrl(productDTO.getImgUrl());
+            // entity.setPrice(productDTO.getPrice());
+            entity = productRepository.save(entity);
+            return new ResponseProductDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundExceptions("Produc not found");
+        }
 
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundExceptions("Produc not found");
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(RequestProductDTO productDTO, Product entity) {
